requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
isodate==0.6.1
lxml==4.9.1
platformdirs==2.5.4
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
urllib3==1.26.13
zeep==4.2.1
